+++ src/ui/control/geolocate_control.test.ts
@@ -55,6 +55,18 @@ describe('GeolocateControl with no options', () => {
         expect(geolocate._geolocateButton.disabled).toBeFalsy();
     });
 
+    test('is disabled when permission is denied and tracking is off', async () => {
+        const geolocate = new GeolocateControl({trackUserLocation: false});
+        map.addControl(geolocate);
+        await sleep(0);
+
+        const click = new window.Event('click');
+        geolocate._geolocateButton.dispatchEvent(click);
+        geolocation.sendError({code: 1, message: 'permission was denied'});
+
+        expect(geolocate._geolocateButton.disabled).toBeTruthy();
+    });
+
     test('has permissions', async () => {
 
         (window.navigator as any).permissions = {